fix(build): move dynamic analytics to client component, fix test cleanup
- next/dynamic with ssr:false not allowed in Server Components (Next.js 16).
  Extract Analytics + SpeedInsights into ClientAnalytics "use client" wrapper.
- Add missing afterEach cleanup to useAnimatedCounter advanced tests, preventing
  unhandled "window is not defined" error from lingering animation frames.
