+++ apps/web/components/ClientAnalytics.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const Analytics = dynamic(
+  () =>
+    import("@vercel/analytics/react").then((m) => ({ default: m.Analytics })),
+  { ssr: false }
+);
+const SpeedInsights = dynamic(
+  () =>
+    import("@vercel/speed-insights/next").then((m) => ({
+      default: m.SpeedInsights,
+    })),
+  { ssr: false }
+);
+
+export function ClientAnalytics() {
+  return (
+    <>
+      <Analytics />
+      <SpeedInsights />
+    </>
+  );
+}
